Move challenge scripts to controllers
